chore: separated common logic within file creations, compile_c, compile_rust and compare_outputs
