+++ crates/intrinsic-test/src/common/mod.rs
@@ -1,4 +1,7 @@
 pub mod cli;
+pub mod compare;
+pub mod gen_c;
+pub mod gen_rust;
 pub mod supporting_test;
 pub mod types;
 pub mod values;
